Refactor `AlertMessage` a bit and add stories
