+++ frontend/src/lib/components/AlertMessage/AlertMessage.scss
@@ -1,6 +1,6 @@
 @import '~/vars';
 
-.lemon-alert-message {
+.AlertMessage {
     background-color: var(--radius-light);
     border-radius: var(--radius);
     padding: 0.5rem 0.75rem;
@@ -9,10 +9,11 @@
     display: flex;
     align-items: center;
 
-    .lemon-alert-message__icon {
+    .AlertMessage__icon {
+        line-height: 0;
         flex-shrink: 0;
         font-size: 1.5rem;
-        margin-right: 0.75rem;
+        margin: 0.25rem 0.75rem 0.25rem 0;
     }
 
     &.info {
