+++ frontend/src/lib/components/AlertMessage/index.ts
@@ -1,0 +1,2 @@
+export { AlertMessage } from './AlertMessage'
+export type { AlertMessageProps } from './AlertMessage'
