--- frontend/src/lib/components/InfoMessage/InfoMessage.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import { AlertMessage, AlertMessageInterface } from './AlertMessage'
-
-/** DEPRECATED: Use `AlertMessage` instead with type = 'info' */
-export function InfoMessage(props: AlertMessageInterface): JSX.Element {
-    return <AlertMessage {...props} />
-}
