+++ frontend/src/scenes/organization/Settings/InviteModal.tsx
@@ -8,9 +8,9 @@ import './InviteModal.scss'
 import { isEmail, pluralize } from 'lib/utils'
 import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
 import { inviteLogic } from './inviteLogic'
-import { InfoMessage } from 'lib/components/InfoMessage/InfoMessage'
 import { IconOpenInNew } from 'lib/components/icons'
 import { LemonButton } from 'lib/components/LemonButton'
+import { AlertMessage } from 'lib/components/AlertMessage'
 
 /** Shuffled placeholder names */
 const PLACEHOLDER_NAMES: string[] = [...Array(10).fill('Jane'), ...Array(10).fill('John'), 'Sonic'].sort(
@@ -20,7 +20,7 @@ const MAX_INVITES_AT_ONCE = 20
 
 export function EmailUnavailableMessage(): JSX.Element {
     return (
-        <InfoMessage style={{ marginTop: 16 }}>
+        <AlertMessage type="info" style={{ marginTop: 16 }}>
             <>
                 This PostHog instance isn't{' '}
                 <a href="https://posthog.com/docs/self-host/configure/email" target="_blank" rel="noopener">
@@ -30,7 +30,7 @@ export function EmailUnavailableMessage(): JSX.Element {
                 .<br />
                 Remember to <u>share the invite link</u> with each team member you invite.
             </>
-        </InfoMessage>
+        </AlertMessage>
     )
 }
 
@@ -153,7 +153,7 @@ export function InviteModal({ visible, onClose }: { visible: boolean; onClose: (
 
                     <div className="mt">
                         {areInvitesCreatable && (
-                            <LemonButton type="primary" onClick={appendInviteRow}>
+                            <LemonButton type="secondary" onClick={appendInviteRow} fullWidth center>
                                 Add another team member
                             </LemonButton>
                         )}
